登場ステータスのフォームを全体のフォームから切り出しました。
